+++ libs/zhi-blog-api/src/index.ts
@@ -20,6 +20,7 @@ import WebPlaceholder from "./lib/WebPlaceholder"
 import CategoryTypeEnum from "./lib/enums/categoryTypeEnum"
 import YamlFormatObj from "./lib/models/yamlFormatObj"
 import YamlConvertAdaptor from "./lib/yamlConvertAdaptor"
+import PreferenceConfig from "./lib/PreferenceConfig"
 
 export { BlogApi, BlogAdaptor }
 export { WebApi, WebAdaptor }
@@ -28,4 +29,5 @@ export { WebConfig, WebPlaceholder }
 export { Post, UserBlog, SiteConfig, CategoryInfo, MediaObject, Attachment }
 export { BlogConstants, BlogTypeEnum, PageTypeEnum, CategoryTypeEnum }
 export { YamlFormatObj, YamlConvertAdaptor }
+export { PreferenceConfig }
 export { type ElectronCookie }