+++ libs/zhi-blog-api/src/lib/models/post.ts
@@ -168,47 +168,18 @@ class Post {
    * @returns {Object} 表示数据的适用于 YAML 的对象
    */
   public toYamlObj(): Record<string, any> {
-    return {
-      /**
-       * 创建日期，已转换为中文格式
-       */
-      created: DateUtil.formatIsoToZh(this.dateCreated.toISOString(), true),
+    const yamlObj: Record<string, any> = {}
 
-      /**
-       * 更新日期，已转换为中文格式。
-       */
-      updated: DateUtil.formatIsoToZh(this.dateUpdated.toISOString(), true),
+    this.dateCreated && (yamlObj.created = DateUtil.formatIsoToZh(this.dateCreated.toISOString(), true))
+    this.dateUpdated && (yamlObj.updated = DateUtil.formatIsoToZh(this.dateUpdated.toISOString(), true))
+    this.title && (yamlObj.title = this.title)
+    this.wp_slug && (yamlObj.slug = this.wp_slug)
+    this.permalink && (yamlObj.permalink = this.permalink)
+    this.shortDesc && (yamlObj.desc = this.shortDesc)
+    this.mt_keywords && (yamlObj.tags = this.mt_keywords?.split(","))
+    this.categories && (yamlObj.categories = this.categories)
 
-      /**
-       * 标题。
-       */
-      title: this.title,
-
-      /**
-       * WordPress 别名
-       */
-      slug: this.wp_slug,
-
-      /**
-       * 永久链接
-       */
-      permalink: this.permalink,
-
-      /**
-       * 简短描述
-       */
-      desc: this.shortDesc,
-
-      /**
-       * 标签
-       */
-      tags: this.mt_keywords,
-
-      /**
-       * 分类列表
-       */
-      categories: this.categories,
-    }
+    return yamlObj
   }
 
   /**
@@ -217,45 +188,19 @@ class Post {
    * @param {Object} yamlObj - 包含要填充对象属性的数据的适用于 YAML 的对象
    */
   public fromYaml(yamlObj: Record<string, any>): void {
-    /**
-     * 创建日期
-     */
-    this.dateCreated = DateUtil.convertStringToDate(yamlObj.created)
-
-    /**
-     * 更新日期
-     */
-    this.dateUpdated = DateUtil.convertStringToDate(yamlObj.updated)
-
-    /**
-     * 标题
-     */
-    this.title = yamlObj.title
-
-    /**
-     * WordPress 别名
-     */
-    this.wp_slug = yamlObj.slug
-
-    /**
-     * 永久链接
-     */
-    this.permalink = yamlObj.permalink
-
-    /**
-     * 简短描述
-     */
-    this.shortDesc = yamlObj.desc
-
-    /**
-     * 标签
-     */
-    this.mt_keywords = yamlObj.tags
-
-    /**
-     * 分类列表
-     */
-    this.categories = yamlObj.categories
+    this.dateCreated = yamlObj?.created ? DateUtil.convertStringToDate(yamlObj.created) : this.dateCreated
+    this.dateUpdated = yamlObj?.updated ? DateUtil.convertStringToDate(yamlObj.updated) : this.dateUpdated
+    this.title = yamlObj?.title ?? this.title
+    this.wp_slug = yamlObj?.slug ?? this.wp_slug
+    this.permalink = yamlObj?.permalink ?? this.permalink
+    this.shortDesc = yamlObj?.desc ?? this.shortDesc
+    // 修复历史遗留问题
+    if (typeof yamlObj?.tags === "string" && yamlObj?.tags.indexOf(",") > -1) {
+      this.mt_keywords = yamlObj?.tag
+    } else {
+      this.mt_keywords = yamlObj?.tags ? yamlObj?.tags.join(",") : this.mt_keywords
+    }
+    this.categories = yamlObj?.categories ?? this.categories
   }
 }