+++ libs/zhi-fetch-middleware/src/lib/commonFetchClient.ts
@@ -25,7 +25,7 @@
 
 import { simpleLogger } from "zhi-lib-base"
 import { BrowserUtil, DeviceDetection, DeviceTypeEnum, SiyuanDevice } from "zhi-device"
-import { StrUtil } from "zhi-common"
+import { JsonUtil, StrUtil } from "zhi-common"
 import { fetchNode } from "./impl/nodeFetch"
 import { fetchChrome } from "./impl/chromeFetch"
 import { fetchMiddleware } from "./impl/middlewareFetch"
@@ -33,9 +33,9 @@ import { fetchMiddleware } from "./impl/middlewareFetch"
 class CommonFetchClient {
   private readonly appInstance: any
 
-  private readonly logger
-  private readonly requestUrl
-  private readonly middlewareUrl
+  private readonly logger: any
+  private readonly requestUrl: string
+  private readonly middlewareUrl: string
 
   constructor(appInstance: any, requestUrl?: string, middlewareUrl?: string, isDev?: boolean) {
     this.appInstance = appInstance
@@ -94,22 +94,28 @@ class CommonFetchClient {
 
     let resJson: any
 
-    const isResponse = response?.status && response?.headers && response?.url
-    const isStream = response?.body instanceof ReadableStream
-    const isString = response?.body instanceof String
+    const isResponse = response?.status !== undefined && response?.headers !== undefined && response?.url !== undefined
+    const isStream = !isResponse && response?.body instanceof ReadableStream
     this.logger.info(`check if response is valid, isResponse=>${isResponse}`)
-    this.logger.info(`check response body is stream =>${isStream}`, typeof response?.body)
-    this.logger.info(`check response body is string =>${isString}`, typeof response?.body)
+    this.logger.info(`check response is stream =>${isStream}`)
 
-    if (!isResponse || !(response instanceof Response) || isStream) {
-      this.logger.debug("response不是Response的实例", typeof response)
+    if (isStream) {
+      this.logger.info("检测到response不是Response的实例，直接返回", typeof response)
       resJson = response
     } else {
       // 解析响应体并返回响应结果
       const statusCode = response.status
-
-      if (statusCode !== 200) {
-        if (statusCode === 401) {
+      const httpSuccess = statusCode >= 200 && statusCode < 300
+      if (!httpSuccess) {
+        this.logger.error("fetch请求错误，response=>", response)
+        let resBody = response?.body
+        if (response?.body instanceof ReadableStream) {
+          resBody = await response.text()
+        }
+        this.logger.error(`fetch请求错误，code ${statusCode}, body =>`, resBody)
+        if (statusCode === 400) {
+          throw new Error("错误请求，服务器不理解请求的语法")
+        } else if (statusCode === 401) {
           throw new Error("因权限不足操作已被禁止")
         } else if (statusCode > 401) {
           if (statusCode === 413) {
@@ -124,26 +130,27 @@ class CommonFetchClient {
           }
           throw new Error(msg)
         } else {
-          throw new Error("fetch请求错误")
+          throw new Error(`fetch请求错误，code is ${statusCode}, body => ${response?.body}`)
         }
       }
 
-      this.logger.debug("isNode=>", BrowserUtil.isNode)
-      this.logger.debug("isElectron=>", BrowserUtil.isElectron())
-      this.logger.debug("isInSiyuanWidget=>", SiyuanDevice.isInSiyuanWidget())
+      this.logger.info("isNode=>", BrowserUtil.isNode)
+      this.logger.info("isElectron=>", BrowserUtil.isElectron())
+      this.logger.info("isInSiyuanWidget=>", SiyuanDevice.isInSiyuanWidget())
       if (BrowserUtil.isNode) {
-        resJson = await response.json()
+        resJson = await this.safeParseBodyJson(response)
       } else if (BrowserUtil.isElectron()) {
-        resJson = await response.json()
+        resJson = await this.safeParseBodyJson(response)
       } else if (SiyuanDevice.isInSiyuanWidget()) {
-        resJson = await response.json()
+        resJson = await this.safeParseBodyJson(response)
       } else {
-        this.logger.debug("解析CORSBody")
-        const corsJson = await response.json()
+        this.logger.debug("开始解析CORSBody")
+        const corsJson = await this.safeParseBodyJson(response)
         resJson = this.parseCORSBody(corsJson)
       }
     }
 
+    this.logger.debug("全部处理完毕，resJson=>", resJson)
     return resJson
   }
 
@@ -154,7 +161,7 @@ class CommonFetchClient {
    * @param middlewareUrl - 可选，当环境不支持时候，必传
    */
   private async fetchEntry(apiUrl: string, fetchOptions: RequestInit, middlewareUrl?: string): Promise<any> {
-    let result
+    let result: any
 
     const deviceType = DeviceDetection.getDevice()
     this.logger.info("deviceType =>", deviceType)
@@ -198,6 +205,15 @@ class CommonFetchClient {
     this.logger.debug("最终返回给前端的数据=>", result)
 
     return result
+  }
+
+  //================================================================
+  // private function
+  //================================================================
+  private async safeParseBodyJson(response: any) {
+    const resText = await response.text()
+    const resJson = JsonUtil.safeParse<any>(resText, {} as any)
+    return resJson
   }
 
   /**