+++ libs/zhi-wechatsync-middleware/src/index.ts
@@ -1,3 +1,0 @@
-import CommonWechatsyncClient from "./lib/commonWechatsyncClient"
-
-export { CommonWechatsyncClient }